backend/analytics/requirements.txt
Flask==2.1.1
pymongo==4.0.1
kafka-python==2.0.2
